change scouting to only happen on creation of a new save
By preloading the ItemChanger data, we can tell if we need to scout or not based on whether ItemChanger is already populated. We have to use the publicized ItemChanger because Load is not public and Load is only trigger on game load, and we need the await to happen earlier.