check for inputs on plugin update, filter out keyboard typing
Fixes bug where some letters would trigger actions in game because of controller support. Moved CheckForInputs to Update because that's when Inputs are handled by other functions in Unity. Added filtering to check for inputs to not trigger actions on input strings, modifier keys that may be used in the Connection Menu, and common keyboard shortcuts. Ctrl-c and ctrl-v now work as copy and paste.