received items as fast as we can, delay notifications instead
This change is definitely a design decision and I'm open to changing it. If someone is loading into a file that has received a lot of items or is playing when a release happens, this choice means they can start playing immediately with what they have received instead of waiting for the receipts one by one.